perf: 分离配置文件
