+++ src/components/flow/designer/config/flow.js
@@ -210,57 +210,3 @@ export const flowConfig = {
     }
   }
 }
-
-// 快捷键
-export const shortcutKeys = {
-  multiple: {
-    code: 16,
-    codeName: 'SHIFT',
-    shortcutName: '多选'
-  },
-  dragContainer: {
-    code: 32,
-    codeName: 'SPACE',
-    shortcutName: '画布拖拽'
-  },
-  dragTool: {
-    code: 68,
-    codeName: 'D',
-    shortcutName: '拖拽工具'
-  },
-  connTool: {
-    code: 76,
-    codeName: 'L',
-    shortcutName: '连线工具'
-  },
-  leftMove: {
-    code: 37,
-    codeName: '←',
-    shortcutName: '左移'
-  },
-  upMove: {
-    code: 38,
-    codeName: '↑',
-    shortcutName: '上移'
-  },
-  rightMove: {
-    code: 39,
-    codeName: '→',
-    shortcutName: '右移'
-  },
-  downMove: {
-    code: 40,
-    codeName: '↓',
-    shortcutName: '下移'
-  },
-  settingModal: {
-    code: 83,
-    codeName: 'CTRL + S',
-    shortcutName: '保存页面'
-  },
-  testModal: {
-    code: 84,
-    codeName: 'CTRL + T',
-    shortcutName: '测试页面'
-  }
-}
