+++ src/components/flow/designer/config/nodes.js
@@ -1,16 +1,3 @@
-export const tools = [
-  {
-    type: 'drag',
-    icon: 'drag',
-    name: '拖拽'
-  },
-  {
-    type: 'connection',
-    icon: 'fork',
-    name: '连线'
-  }
-]
-
 export const commonNodes = [
   {
     type: 'start',
