+++ src/components/flow/designer/config/tools.js
@@ -1,0 +1,12 @@
+export const tools = [
+  {
+    type: 'drag',
+    icon: 'drag',
+    name: '拖拽'
+  },
+  {
+    type: 'connection',
+    icon: 'fork',
+    name: '连线'
+  }
+]
